Move some model-related modules to dl_model_tools
