+++ lib/dl_api_client/pyproject.toml
@@ -14,7 +14,6 @@ attrs = ">=22.2.0"
 marshmallow = ">=3.19.0"
 marshmallow-oneofschema = ">=3.0.1"
 python = ">=3.10, <3.13"
-datalens-utils = {path = "../dl_utils"}
 datalens-constants = {path = "../dl_constants"}
 datalens-model-tools = {path = "../dl_model_tools"}
