+++ lib/dl_api_lib/dl_api_lib/schemas/action.py
@@ -34,10 +34,6 @@ from dl_api_lib.request_model.data import (
 )
 from dl_api_lib.schemas.filter import ObligatoryFilterSchema
 from dl_api_lib.schemas.parameters import ParameterValueConstraintSchema
-from dl_api_lib.schemas.values import (
-    ValueSchema,
-    WithNestedValueSchema,
-)
 from dl_constants.enums import (
     AggregationFunction,
     CalcMode,
@@ -46,6 +42,10 @@ from dl_constants.enums import (
 from dl_model_tools.schema.base import (
     BaseSchema,
     DefaultValidateSchema,
+)
+from dl_model_tools.schema.typed_values import (
+    ValueSchema,
+    WithNestedValueSchema,
 )
