+++ lib/dl_api_lib/dl_api_lib/schemas/typed_query.py
@@ -1,20 +1,20 @@
 import attr
 from marshmallow import fields as ma_fields
 
-from dl_api_lib.schemas.values import (
-    ValueSchema,
-    WithNestedValueSchema,
-)
 from dl_constants.enums import (
     DashSQLQueryType,
     UserDataType,
 )
-from dl_core.values import BIValue
 from dl_model_tools.schema.base import (
     BaseSchema,
     DefaultSchema,
 )
 from dl_model_tools.schema.dynamic_enum_field import DynamicEnumField
+from dl_model_tools.schema.typed_values import (
+    ValueSchema,
+    WithNestedValueSchema,
+)
+from dl_model_tools.typed_values import BIValue
 
 
 @attr.s(frozen=True)
