+++ lib/dl_core/dl_core/fields.py
@@ -30,7 +30,7 @@ from dl_constants.enums import (
 )
 from dl_core.components.ids import FieldId
 from dl_core.exc import FieldNotFound
-from dl_core.values import BIValue
+from dl_model_tools.typed_values import BIValue
 
 
 LOGGER = logging.getLogger(__name__)
