+++ lib/dl_core/dl_core/us_manager/storage_schemas/dataset.py
@@ -56,7 +56,7 @@ from dl_core.us_dataset import Dataset
 from dl_core.us_manager.storage_schemas.base import DefaultStorageSchema
 from dl_core.us_manager.storage_schemas.data_source_collection import GenericDataSourceCollectionStorageSchema
 from dl_core.us_manager.storage_schemas.error_registry import ComponentErrorListSchema
-from dl_core.values import (
+from dl_model_tools.typed_values import (
     ArrayFloatValue,
     ArrayIntValue,
     ArrayStrValue,
