+++ lib/dl_model_tools/dl_model_tools/typed_values.py
