+++ lib/dl_model_tools/pyproject.toml
@@ -14,6 +14,7 @@ attrs = ">=22.2.0"
 datalens-constants = {path = "../dl_constants"}
 dynamic-enum = {path = "../dynamic_enum"}
 marshmallow = ">=3.19.0"
+marshmallow-oneofschema = ">=3.0.1"
 python = ">=3.10, <3.13"
 
 [tool.poetry.group.tests.dependencies]
